airflow/requirements.txt
psycopg2-binary==2.9.7
pandas==2.0.3
requests==2.31.0
boto3==1.28.0
python-dotenv==1.1.1